requirements.txt
numpy==1.21.5
lightgbm==3.3.2
sklearn-genetic==0.5.1
scikit-learn==1.1.2
pyyaml==6.0